Move test_profile into test_model
