+++ .github/workflows/tests.yml
@@ -78,7 +78,6 @@ jobs:
             pymc/tests/test_model_graph.py
             pymc/tests/ode/test_ode.py
             pymc/tests/ode/test_utils.py
-            pymc/tests/test_profile.py
             pymc/tests/step_methods/hmc/test_quadpotential.py
 
       fail-fast: false
@@ -147,7 +146,7 @@ jobs:
         python-version: ["3.8"]
         test-subset:
           - pymc/tests/test_variational_inference.py pymc/tests/test_initial_point.py
-          - pymc/tests/test_model.py pymc/tests/test_profile.py pymc/tests/test_step.py
+          - pymc/tests/test_model.py pymc/tests/test_step.py
           - pymc/tests/gp/test_cov.py pymc/tests/gp/test_gp.py pymc/tests/gp/test_mean.py pymc/tests/gp/test_util.py pymc/tests/ode/test_ode.py pymc/tests/ode/test_utils.py pymc/tests/test_smc.py pymc/tests/test_parallel_sampling.py
           - pymc/tests/test_sampling.py pymc/tests/test_posteriors.py
