+++ pymc/tests/test_model.py
@@ -1118,3 +1118,19 @@ def test_mixed_contexts():
         pm.Model.get_context(error_if_none=True)
     with pytest.raises((ValueError, TypeError)):
         modelcontext(None)
+
+
+class TestProfile:
+    def setup_method(self):
+        _, self.model, _ = simple_model()
+
+    def test_profile_model(self):
+        assert self.model.profile(self.model.logp()).fct_call_time > 0
+
+    def test_profile_variable(self):
+        rv = self.model.basic_RVs[0]
+        assert self.model.profile(self.model.logp(vars=[rv], sum=False)).fct_call_time
+
+    def test_profile_count(self):
+        count = 1005
+        assert self.model.profile(self.model.logp(), n=count).fct_callcount == count
